changed CPU count from float to int to work with multiprocessors

+++ utils/ASOget.py
@@ -97,7 +97,7 @@ else:
     
 #set multiprocessing limits
 CPUS = len(os.sched_getaffinity(0))
-CPUS = (CPUS/2)-2
+CPUS = int((CPUS/2)-2)
 
 
 class ASODataTool:
